Fix search-paths when loaded graphically on macOS
I recently replaced the loader-script used by http://emacsformacosx.com/
with a symlink to the actual Emacs.app binary. This was done to solve an
issue with permissions on macOS (caldwell/build-emacs#84), but this also
exposed an issue with $PATH handling (caldwell/build-emacs#39).

I don't want Ruby scripts to have unrestricted filesystem access, nor do
I want userland folders to be inaccessible via dired. To workaround both
issues, Emacs.app now spawns a shell at startup in order to reread $PATH
(and $MANPATH) so that commands in ~/.files/bin` or wherever are visible
to a graphical Emacs session.